refactor(package): reorganise all packages based on feature
