--- src/main/java/nextstep/subway/application/exception/ConstraintErrorAdvice.java
@@ -1,29 +1,0 @@
-package nextstep.subway.application.exception;
-
-import org.aspectj.weaver.ast.Not;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.context.request.WebRequest;
-
-@ControllerAdvice
-public class ConstraintErrorAdvice {
-    @ExceptionHandler(Exception.class)
-    public final ResponseEntity<String> handleAllExceptions(Exception ex, WebRequest request) {
-        return ResponseEntity.internalServerError().body(ex.getMessage());
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException ex) {
-        return ResponseEntity.badRequest().body(ex.getMessage());
-    }
-
-    @ExceptionHandler(NotFoundException.class)
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ResponseEntity<Object> handleNotFoundException(NotFoundException ex) {
-        return ResponseEntity.notFound().build();
-    }
-}
