--- src/main/java/nextstep/subway/application/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package nextstep.subway.application.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
