--- src/main/java/nextstep/subway/domain/line/LineRepository.java
@@ -1,6 +1,0 @@
-package nextstep.subway.domain.line;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LineRepository extends JpaRepository<Line, Long> {
-}
