+++ src/main/java/nextstep/subway/line/Line.java
@@ -1,0 +1,71 @@
+package nextstep.subway.line;
+
+import javax.persistence.Embedded;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import nextstep.subway.station.Station;
+
+@Entity
+public class Line {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Embedded
+    private LineName name;
+
+    @Embedded
+    private LineColor color;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "up_station_id", nullable = false)
+    private Station upStation;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "down_station_id", nullable = false)
+    private Station downStation;
+
+    @Embedded
+    private LineDistance distance;
+
+    protected Line() {
+    }
+
+    public Line(String name, String color, Station upStation, Station downStation, int distance) {
+        this.name = new LineName(name);
+        this.color = new LineColor(color);
+        this.upStation = upStation;
+        this.downStation = downStation;
+        this.distance = new LineDistance(distance);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name.value();
+    }
+
+    public String getColor() {
+        return color.value();
+    }
+
+    public Station getUpStation() {
+        return upStation;
+    }
+
+    public Station getDownStation() {
+        return downStation;
+    }
+
+    public void update(String name, String color) {
+        this.name = new LineName(name);
+        this.color = new LineColor(color);
+    }
+}
