+++ src/main/java/nextstep/subway/line/LineRepository.java
@@ -1,0 +1,6 @@
+package nextstep.subway.line;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LineRepository extends JpaRepository<Line, Long> {
+}
