+++ src/main/java/nextstep/subway/shared/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package nextstep.subway.shared.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
