+++ src/main/java/nextstep/subway/station/StationController.java
@@ -1,0 +1,42 @@
+package nextstep.subway.station;
+
+import nextstep.subway.station.dto.StationRequestDto;
+import nextstep.subway.station.dto.StationResponseDto;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class StationController {
+    private StationService stationService;
+
+    public StationController(StationService stationService) {
+        this.stationService = stationService;
+    }
+
+    @PostMapping("/stations")
+    public ResponseEntity<StationResponseDto> createStation(@RequestBody StationRequestDto stationRequest) {
+        StationResponseDto station = stationService.saveStation(stationRequest);
+        return ResponseEntity.created(URI.create("/stations/" + station.getId())).body(station);
+    }
+
+    @GetMapping(value = "/stations", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<StationResponseDto>> showStations() {
+        return ResponseEntity.ok().body(stationService.findAllStations());
+    }
+
+    @DeleteMapping("/stations/{id}")
+    public ResponseEntity deleteStation(@PathVariable Long id) {
+        stationService.deleteStationById(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @ExceptionHandler(DataIntegrityViolationException.class)
+    public ResponseEntity handleIllegalArgsException() {
+        return ResponseEntity.badRequest().build();
+    }
+}
