+++ src/main/java/nextstep/subway/station/StationRepository.java
@@ -1,0 +1,10 @@
+package nextstep.subway.station;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface StationRepository extends JpaRepository<Station, Long> {
+    @Override
+    List<Station> findAll();
+}
