+++ src/main/java/nextstep/subway/station/dto/StationRequestDto.java
@@ -1,0 +1,21 @@
+package nextstep.subway.station.dto;
+
+import nextstep.subway.station.Station;
+
+public class StationRequestDto {
+    private String name;
+
+    private StationRequestDto() {}
+
+    public StationRequestDto(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Station toStation() {
+        return new Station(name);
+    }
+}
