--- src/main/java/nextstep/subway/ui/LineController.java
@@ -1,55 +1,0 @@
-package nextstep.subway.ui;
-
-import java.net.URI;
-import java.util.List;
-import nextstep.subway.application.LineService;
-import nextstep.subway.dto.line.LineRequest;
-import nextstep.subway.dto.line.LineResponse;
-import nextstep.subway.dto.line.UpdateLineRequest;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class LineController {
-    private final LineService lineService;
-
-    public LineController(LineService lineService) {
-        this.lineService = lineService;
-    }
-
-    @PostMapping(value = "/lines")
-    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
-        LineResponse line = lineService.createLine(lineRequest);
-        return ResponseEntity.created(URI.create("/lines/" + line.getId())).body(line);
-    }
-
-    @GetMapping(value = "/lines", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<LineResponse>> showLines() {
-        return ResponseEntity.ok().body(lineService.findAllLines());
-    }
-
-    @GetMapping(value = "/lines/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LineResponse> showLine(@PathVariable("id") Long id) {
-        return ResponseEntity.ok().body(lineService.findLine(id));
-    }
-
-    @PutMapping(value = "/lines/{id}")
-    public ResponseEntity<LineResponse> updateLine(@PathVariable("id") Long id,
-                                                   @RequestBody UpdateLineRequest updateLineRequest) {
-        lineService.updateLine(id, updateLineRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping(value = "/lines/{id}")
-    public ResponseEntity<LineResponse> deleteLine(@PathVariable("id") Long id) {
-        lineService.deleteLine(id);
-        return ResponseEntity.noContent().build();
-    }
-}
